Let maybe-pod-name-at-point work independently of current buffer
`kubernetes-utils-maybe-pod-name-at-point` can be used to retrieve the
current pod name at point while in the kubernetes-overview buffer.

In certain cases it is useful to be able to access this information
while a different buffer is active.

This commit refactors the function to instead use a temporary current
buffer binding to always perform the lookup in the overview buffer.