Match all selector labels instead of only "name" (#103)
Kubernetes pods can match a deployment by any label. The current code
assumes that there's only 1 label, and the label is always called
"name". That's not actually true, resulting in all (or no) pods being
listed when unfolding any deployment.

This commit applies all `matchLabels` in a deployment to all pods,
getting a more accurate list of which pods belong to which deployment.