AST: Properly render lists w/ dash-prefixed elements
Closes #81.

Current AST rendering logic for bullet lists was such that lists that
we expect to be rendered like so:
```yaml
args:
  - --foo
  - --bar
  - --baz
```

Were instead getting rendered along the lines of the following:
```yaml
args:
  --foo--bar--baz
```

This PR fixes rendering of bullet lists to align with the former
output.

Incidentally, we also fix up a related bug in list rendering that
results in the following expected output:
```yaml
- foo
- -bar
- --baz
```

Instead getting rendered like so:
```yaml
- foo
-bar
--baz
```

Both of these scenarios can be confirmed by checking out the
contributed test cases, less the changes to
`kubernetes-{ast,yaml}.el`, and running them on `master`.