Genericize and consolidate `kubectl delete` operations (#198)
Relates to #69.

Currently, we define one function for `delete`-ing all entities of a given resource type, for each of a finite set of resources, e.g. `k8s-kubectl-delete-{pods, nodes, ...}`. All of these functions are identical save for the resource type that they retrieve.

As such, we generalize to a single implementation—`k8s-kubectl-delete`—that accepts the resource type name and performs the `delete` accordingly.

We additionally clean up unit tests and, finally, remove the granular, per-type `get` functions. 

In particular, with regards to unit tests, we observe that, once the resource type name is abstracted away, there's no practical reason to have one test case for each resource type. As such, rather than consolidate all per-type test cases into a single test function, we simply keep one case for each of the two scenarios "delete success" and "delete fail," and remove the rest of the test functions as superfluous.