Start a changelog
This PR introduces a changelog practice to the project.

Changelogs are a helpful way for users to be made aware of changes to the
package from version to version. It does not replace GitHub's
[releases](https://github.com/kubernetes-el/kubernetes-el/releases) page for the
project, but rather supplements it. Package maintainers can add entries to the
changelog's "Unreleased" section as we implement changes, copying over the
contents into a new version's section and notes when we cut a new version.

Details of this process are explained further in
[keepachangelog.com](https://keepachangelog.com/en/1.0.0/). We base this
changelog on the aforementioned methodology.