Tests: Convert subset of AST-rendering tests to Buttercup (#225)
In response to #224, here we convert a subset of the existing AST parsing tests from ERT to Buttercup. These tests were selected for proactive conversion here due to their simplicity; they by and large follow the common pattern of **rendering an input AST** and **asserting on the rendered string**.

We also update CI to combine test coverage for both the ERT and Buttercup tests.

With this PR, we enter a state in which we have both Buttercup and ERT tests co-existing. This is arguably okay, and it doesn't seem particularly valuable to proactively convert all tests to Buttercup. However, it is my recommendation that new tests be written in Buttercup, as it is IMO vastly more user-friendly and flexible (see #224 for more details).