Extract generic form of resource force-refresh logic (#229)
Relates to #69.

This PR extracts the core of the resource-specific
`kubernetes-<resource>-refresh-now` functions into the resource-generic
`kubernetes-state--refresh-now` function. This function represents the "core" of
the function template used to generate the resource-specific force-refresh
functions.

This PR leaves the resource-specific refresh-now functions in place,
re-implementing them as wrappers around the new resource-generic version. The
resource-specific functions have test coverage, which provides a vote of
confidence as to the safety of this change. We'll leave sunsetting of the
resource-specific functions' use to subsequent PRs.