Process tracking: Implement new, comprehensive, agnostic tracker
Relates to #234.

As the first step towards a resource-agnostic process tracking strategy, this PR
implements `kubernetes--process-ledger`, an EIEIO-based ledger class to keep
track of processes corresponding to given resources. This PR also reimplements
the existing resource-specific process-related constructs,
e.g. `kubernetes-process-pods-process-live-p`, to be with respect to a global
instance of this new ledger class.