Implement initialization + tracking of kubectl proxy process
Contributes to #235.

This PR introduces the `kubernetes--with-proxy` macro, which allows for
executing a block of code with the guarantee of a `kubectl proxy` kicking off,
or having already kicked off, in the background. This paves the way for us to
make HTTP requests to the Kubernetes API via the proxy.

One downside here is that the proxy process lives indefinitely once it's kicked
off. Ideally, this process would have a TTL where it shuts down after going
unused for a specified amount of time. This can be future work. Due to this,
however, we should only merge this PR in once #232 is addressed.