feat(configmaps): show (truncated) data in overview (#199)
Shows configmap data in a new section in the overview.

Both keys and values are truncated. If the value spans multiple lines,
only the first line is shown (and maybe truncated). To show this
truncation, a forward slash (dimmed) is shown.

Adds a new case to the configmap test response; the expected case is
updated accordingly.