Expand documentation w/ tutorials + "getting started" section (#250)
* Expand documentation w/ tutorials + "getting started" section

This PR adds a dedicated "getting started" section to give users introductory
overview into: what the package is; how to get it; how to start it; etc. It also
lays the foundation for a dedicated section of tutorials to walk users through a
representative set of targeted high-level use cases. It also creates a new
"how-to" section to explain how to perform specific, atomic operations,
e.g. change namespace/context and the like.

Some sections are left bare with TODOs for later expansion.

* Update links

* update more links

* Update readme

* Hide meaningless in home

* Use tables for keybindings/functions

* fix typo

* Add context renaming how-to