Introduce primitives for tracking processes w/ ports + addresses (#256)
Contributes to #247.

This PR introduces `kubernetes--ported-process-record`, a class that facilitates
interaction with processes that correspond to local ports. It provides methods
such as:

- Interpolating the base URL, e.g. `http://localhost:<port>`;

- Waiting for response code from a specific endpoint thereon,
e.g. `http://localhost:<port>/livez` (for the Kubernetes server);

This class will be necessary for systematic tracking of proxy servers created by
`kubectl` as outlined in #247 and implemented #252.