Move `save-window-state` fn from `k8s-util` to `k8s-core` (#278)
Closes #265.

The function `kubernetes-utils--save-window-state` has created an implicit
dependency cycle between `kubernetes-core.el` and `kubernetes-utils.el`. This is
not outright caused by the former `require`-ing the latter, which is itself a
mistake; adding the `require` exposes the import cycle.

This symptom is an excellent example of the overly ambiguous and broad scope of
the `-utils` package. Since the `--save-window-state` function is arguably
"core" functionality, we simply move it to the `-core` package and rename all
references accordingly.