Enable discovery of API groups from cluster (#302)
Contributes to #235.

This PR adds the `kubernetes--get-all-groups` function. This queries the Kubernetes
API via HTTP proxy to discover the API groups present on the cluster.

At the same time, we also introduce a collection of macros to facilitate
interaction with `kubectl` proxy servers, e.g. lifecycle management.

This logic represents the first step to discovering all resource kinds, custom
or otherwise, on the given cluster.